+++ src/main/java/com/misbah/kaoutar/misbahkaoutarexamjee/repositories/AgenceRepository.java
@@ -3,5 +3,5 @@ package com.misbah.kaoutar.misbahkaoutarexamjee.repositories;
 import com.misbah.kaoutar.misbahkaoutarexamjee.entities.Agence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgenceRepository extends JpaRepository<Agence, Integer> {
+public interface AgenceRepository extends JpaRepository<Agence, Long> {
 }
