+++ src/main/java/com/misbah/kaoutar/misbahkaoutarexamjee/repositories/MotoRepository.java
@@ -3,5 +3,5 @@ package com.misbah.kaoutar.misbahkaoutarexamjee.repositories;
 import com.misbah.kaoutar.misbahkaoutarexamjee.entities.Moto;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MotoRepository extends CrudRepository<Moto, Integer> {
+public interface MotoRepository extends CrudRepository<Moto, Long> {
 }
