+++ src/main/java/com/misbah/kaoutar/misbahkaoutarexamjee/repositories/VoitureRepository.java
@@ -3,5 +3,5 @@ package com.misbah.kaoutar.misbahkaoutarexamjee.repositories;
 import com.misbah.kaoutar.misbahkaoutarexamjee.entities.Voiture;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VoitureRepository extends CrudRepository<Voiture, Integer> {
+public interface VoitureRepository extends CrudRepository<Voiture, Long> {
 }
